Add objective C `expr` methods.

+++ ReQL-expr.c
@@ -15,7 +15,7 @@ _ReQL_Op_t *_reql_expr_number(double val) {
   return term;
 }
 
-_ReQL_Op_t *_reql_expr_string(const char *val, unsigned int str_len) {
+_ReQL_Op_t *_reql_expr_string(const char *val, unsigned long str_len) {
   _ReQL_Op_t *term = _reql_expr_null();
   term->dt = _REQL_R_STR;
   term->str = val;
