+++ ReQL-expr.h
@@ -9,7 +9,7 @@ struct _ReQL_Op_s {
   int tt;
   int dt;
 
-  unsigned int str_len;
+  unsigned long str_len;
   const char *str;
 
   double num;
@@ -26,7 +26,7 @@ typedef struct _ReQL_Op_s _ReQL_Op_t;
 _ReQL_Op_t *_reql_expr_bool(int val);
 _ReQL_Op_t *_reql_expr_null();
 _ReQL_Op_t *_reql_expr_number(double val);
-_ReQL_Op_t *_reql_expr_string(const char *val, unsigned int str_len);
+_ReQL_Op_t *_reql_expr_string(const char *val, unsigned long str_len);
 
 _ReQL_Op_t *_reql_expr_array();
 void _reql_array_append(_ReQL_Op_t *arr, _ReQL_Op_t *val);
